\name{Assignment3doc}
\alias{Assignment3doc}

\title{
Assignment 3
}
\description{
2 functions for assignment 3
}
\usage{
points_and_line_graph(x, y, title, x_label, y_label)
simulate(runs, single_run_function, input_value)
}

\arguments{
  \item{points_and_line_graph}{
  A preset scatterplot with regression line
}
  \item{simulate}{
  runs several simulations of a certain funtion that is created
}
}

\author{
Tian Lin
}
